WIP(RouteCardData): Initial static data scaffold
